ARM: ux500: prcmu-qos-debug: Fix warning
ST-Ericsson Linux next: Not tested, ask SSM for ER
ST-Ericsson ID: -
ST-Ericsson FOSS-OUT ID: Trivial

Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/19583

+++ arch/arm/mach-ux500/prcmu-debug.c
@@ -299,7 +299,7 @@ static ssize_t opp_write(struct file *file,
 
 	prcmu_qos_force_opp(prcmu_qos_class, i);
 
-	pr_info("prcmu debug: forced OPP for %d to %d\n", prcmu_qos_class, i);
+	pr_info("prcmu debug: forced OPP for %d to %d\n", prcmu_qos_class, (int)i);
 
 	return buf_size;
 }
